Refactor client
We were very confused as to why our solution wasn't working, so
Utkarsh sat down with us to hold our hand through much of this.

Our old way of structuring the client was broken in a few ways. First,
debug statements had to be removed for the sake of Bats, but that is
only a superficial change.

One of the biggest changes was taking out a lot of the construction of
threads and the to/from classes from main and putting it into a start()
that is called by main on a default EchoClient() constructor. We
think this allows EchoClient to create one proper thread that spawns
Runnable threads in the to/from's.

Another change was that rather than Input/Output Streams being static
fields, they are made locally in To/From ServerThread methods. So
instead of them being constructed in main, they are the result of
the to/from constructor's being passed the socket, with the sockets
being locally turned into input/output streams.

We moved flushing out of the while loops.

Co-authored-by: UtkarshKr077 <kumar375@morris.umn.edu>